requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
Pillow==9.0.0
pypng==0.0.21
PyQRCode==1.2.1
pyshorteners==1.0.1
pytz==2021.3
pyzbar==0.1.8
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
